로그인 권한 확인 알림, 텍스트 크기 조절, 모달 오류 수정, confirm 추가

+++ client/components/yeonwoo/BtnDropdown.tsx
@@ -1,7 +1,7 @@
 /*
  * 책임 작성자: 박연우
  * 최초 작성일: 2022-11-14
- * 최근 수정일: 2022-12-03
+ * 최근 수정일: 2022-12-05
  */
 
 import { faUser } from '@fortawesome/free-regular-svg-icons';
@@ -26,16 +26,22 @@ export const BtnDropdown = () => {
   const router = useRouter();
 
   const onClickLogout = async () => {
-    await axios.delete('/api/auth/token', {
-      headers: {
-        RefreshToken: localStorage.getItem('refreshToken'),
-        'ngrok-skip-browser-warning': '111',
-      },
-    });
-    localStorage.clear();
-    setIsLogin(false);
-    setDataHeader(null);
-    router.push('/');
+    if (window.confirm('로그아웃 하시겠습니까?')) {
+      try {
+        await axios.delete('/api/auth/token', {
+          headers: {
+            RefreshToken: localStorage.getItem('refreshToken'),
+            'ngrok-skip-browser-warning': '111',
+          },
+        });
+        localStorage.clear();
+        setIsLogin(false);
+        setDataHeader(null);
+        router.push('/');
+      } catch (error) {
+        alert('에러발생 : 다시 시도 부탁드립니다');
+      }
+    }
   };
 
   const getPoints = () => {
@@ -56,18 +62,18 @@ export const BtnDropdown = () => {
           <ul className="border border-solid border-black border-opacity-10 border-spacing-1 right-0 min-w-[200px] rounded-xl absolute top-8 bg-background-gray z-20">
             <li className="pt-4 pb-1 mx-4 flex justify-between items-center border-b border-solid">
               <span className="text-xs">나의 모락</span>
-              <span className="text-lg font-semibold">{`✨ ${points} 모락`}</span>
+              <span className="text-sm font-semibold">{`✨ ${points} 모락`}</span>
             </li>
             <Link href={`/dashboard/${localStorage.getItem('userId')}`}>
               <li className="hover:bg-main-yellow hover:bg-opacity-40 hover:cursor-pointer mt-2 py-1 px-4 rounded-xl text-[15px]">
                 <FontAwesomeIcon icon={faUser} size="sm" />
-                <span className="ml-2">대시보드</span>
+                <span className="ml-2 text-sm">대시보드</span>
               </li>
             </Link>
             <Link href="/edit-profile">
               <li className="hover:bg-main-yellow hover:bg-opacity-40 hover:cursor-pointer py-1 mb-2 px-4 rounded-xl text-[15px]">
                 <FontAwesomeIcon icon={faUser} size="sm" />
-                <span className="ml-2">개인정보 수정</span>
+                <span className="ml-2 text-sm">개인정보 수정</span>
               </li>
             </Link>
             <li
@@ -75,7 +81,7 @@ export const BtnDropdown = () => {
               onClick={onClickLogout}
             >
               <FontAwesomeIcon icon={faArrowRightFromBracket} size="sm" />
-              <span className="ml-2">로그아웃</span>
+              <span className="ml-2 text-sm">로그아웃</span>
             </li>
           </ul>
         </div>
